Fix a few typos in basic.py add missing docstring.
Decided to fix a few typos in basic.py while I was reading it. The globals update function was missing a docstring. Not sure if it needs a docstring since in this case it's an empty wrapper but the other update functions (node/edge) were also empty wrappers and had docstrings.